+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1286,7 +1286,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-22T09:24:31.787Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-22T13:07:03.335Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1338,7 +1338,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-22T09:24:31.787Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-22T13:07:03.335Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1390,7 +1390,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-22T09:24:31.786Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-22T13:07:03.335Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1464,5 +1464,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-22T09:24:21.985Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-22T13:06:53.427Z")
     r"""The date and time the resource was last updated"""