+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-22T09:24:21.985Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-22T13:06:53.427Z")
     r"""The date and time the resource was last updated"""