+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K85KQYYH5J2Q1BVGVSJN6SMS"
+        "01K860FDJNEZVEHGT1SJZ080F5"
     )
     r"""The id of the resource"""